--- a/v3/go.mod
+++ b/v3/go.mod
@@ -1,3 +1,3 @@
 module github.com/psilva261/timsort/v3
 
-go 1.14
+go 1.22