Moving to generics to make more type safe and possible faster given the type sorting against.

The changes are in `https://github.com/psilva261/timsort/commit/dcbad9d8eb6ee57292f3b9ac863c01eb3b376d86`

Benchmark diff ( on laptop so result may vary with better cooling)
```bash
goos: linux
goarch: amd64
pkg: github.com/psilva261/timsort/v2
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                               │  v2/bench.txt  │              v3/bench.txt               │
                               │     sec/op     │     sec/op      vs base                 │
TimsortXor100-20                  2.779µ ±  17%    2.328µ ±  11%   -16.21% (p=0.000 n=10)
TimsortInterXor100-20             4.328µ ±  56%    3.282µ ±   8%         ~ (p=0.089 n=10)
StandardSortXor100-20             3.838µ ±  73%    3.254µ ±  13%   -15.22% (p=0.011 n=10)
TimsortSorted100-20               711.6n ±   9%    823.4n ±  64%   +15.72% (p=0.011 n=10)
TimsortInterSorted100-20          1.105µ ±  42%    1.028µ ±  10%    -6.88% (p=0.019 n=10)
StandardSortSorted100-20          409.9n ±  16%    412.9n ±   4%         ~ (p=0.912 n=10)
TimsortRevSorted100-20            763.7n ±  24%    794.9n ±  45%         ~ (p=0.529 n=10)
TimsortInterRevSorted100-20       1.207µ ±   5%    1.200µ ±  10%         ~ (p=0.725 n=10)
StandardSortRevSorted100-20       4.011µ ±  15%    4.497µ ±  72%         ~ (p=0.052 n=10)
TimsortRandom100-20               3.073µ ±  15%    4.466µ ±  29%   +45.33% (p=0.011 n=10)
TimsortInterRandom100-20          3.468µ ±   8%    3.574µ ±  14%         ~ (p=0.393 n=10)
StandardSortRandom100-20          6.088µ ± 109%    4.745µ ±   7%   -22.06% (p=0.001 n=10)
TimsortXor1K-20                  111.16µ ± 108%    50.79µ ±  23%   -54.31% (p=0.000 n=10)
TimsortInterXor1K-20             142.73µ ± 163%    43.33µ ±  73%   -69.65% (p=0.000 n=10)
StandardSortXor1K-20             140.44µ ±  68%    58.26µ ±  17%   -58.51% (p=0.000 n=10)
TimsortSorted1K-20                6.136µ ±  27%    4.882µ ±  47%         ~ (p=0.143 n=10)
TimsortInterSorted1K-20           12.45µ ±  39%    11.23µ ±  29%         ~ (p=0.481 n=10)
StandardSortSorted1K-20           3.704µ ±  14%    6.704µ ±  24%   +80.98% (p=0.001 n=10)
TimsortRevSorted1K-20             7.290µ ±  15%    5.876µ ±  37%         ~ (p=0.075 n=10)
TimsortInterRevSorted1K-20        15.82µ ± 104%    13.70µ ±  23%   -13.40% (p=0.019 n=10)
StandardSortRevSorted1K-20       132.06µ ± 131%    47.74µ ±  24%   -63.85% (p=0.000 n=10)
TimsortRandom1K-20               117.42µ ±  38%    64.75µ ±  61%   -44.86% (p=0.000 n=10)
TimsortInterRandom1K-20          116.78µ ±  35%    81.27µ ±  55%   -30.41% (p=0.011 n=10)
StandardSortRandom1K-20           119.3µ ±  28%    117.7µ ±  47%         ~ (p=0.796 n=10)
TimsortXor1M-20                   61.92m ±  14%    45.19m ±   5%   -27.02% (p=0.000 n=10)
TimsortInterXor1M-20              63.45m ±   5%    67.81m ±  13%         ~ (p=0.143 n=10)
StandardSortXor1M-20              128.8m ±   7%    127.4m ±   6%         ~ (p=0.631 n=10)
TimsortSorted1M-20                3.182m ±  14%    3.375m ±  47%         ~ (p=0.796 n=10)
TimsortInterSorted1M-20           7.081m ±  52%    5.909m ±  62%         ~ (p=0.165 n=10)
StandardSortSorted1M-20           6.138m ±  39%   12.584m ±  20%  +105.03% (p=0.000 n=10)
TimsortRevSorted1M-20             9.803m ± 131%    6.557m ± 127%         ~ (p=0.075 n=10)
TimsortInterRevSorted1M-20        11.20m ±  14%    10.96m ±  22%         ~ (p=0.579 n=10)
StandardSortRevSorted1M-20        83.70m ±  27%    86.26m ±  32%         ~ (p=0.739 n=10)
TimsortRandom1M-20                228.3m ±  18%    160.1m ±   4%   -29.86% (p=0.000 n=10)
TimsortInterRandom1M-20           275.7m ±  46%    254.2m ±   6%    -7.79% (p=0.019 n=10)
StandardSortRandom1M-20           437.0m ±  49%    433.2m ±  28%         ~ (p=0.529 n=10)
TimsortIXor100-20                 2.590µ ±   8%    2.093µ ±  13%   -19.19% (p=0.000 n=10)
StandardSortIXor100-20            3.357µ ±  54%    1.933µ ±  13%   -42.41% (p=0.000 n=10)
TimsortISorted100-20              925.5n ±  68%    551.6n ±  92%   -40.40% (p=0.005 n=10)
StandardSortISorted100-20         402.9n ±  82%    280.0n ± 150%         ~ (p=0.075 n=10)
TimsortIRevSorted100-20           894.5n ±  19%    542.3n ±   6%   -39.37% (p=0.001 n=10)
StandardSortIRevSorted100-20      578.2n ±  81%    391.6n ±  92%         ~ (p=0.063 n=10)
TimsortIRandom100-20              2.277µ ±  35%    3.735µ ±  62%         ~ (p=0.052 n=10)
StandardSortIRandom100-20         3.437µ ±  34%    2.068µ ±   3%   -39.83% (p=0.000 n=10)
TimsortIXor1K-20                  25.61µ ±  22%    24.90µ ±  31%         ~ (p=0.579 n=10)
StandardSortIXor1K-20             40.79µ ±  24%    45.63µ ±  45%         ~ (p=0.436 n=10)
TimsortISorted1K-20               2.829µ ±  29%    2.992µ ±  21%         ~ (p=1.000 n=10)
StandardSortISorted1K-20          1.872µ ±  13%    2.043µ ±  11%         ~ (p=0.165 n=10)
TimsortIRevSorted1K-20            3.408µ ±  27%    7.423µ ± 111%  +117.84% (p=0.000 n=10)
StandardSortIRevSorted1K-20       6.007µ ±  34%    2.936µ ±  61%   -51.13% (p=0.009 n=10)
TimsortIRandom1K-20              137.42µ ± 231%    47.95µ ±  17%   -65.11% (p=0.000 n=10)
StandardSortIRandom1K-20         140.14µ ± 144%    60.36µ ±   8%   -56.93% (p=0.000 n=10)
TimsortIXor1M-20                  57.35m ±   5%    38.04m ±   7%   -33.68% (p=0.000 n=10)
StandardSortIXor1M-20             68.68m ±  34%    45.04m ±   3%   -34.41% (p=0.001 n=10)
TimsortISorted1M-20               2.359m ±  35%    2.553m ±  30%         ~ (p=0.393 n=10)
StandardSortISorted1M-20          2.754m ±  31%    1.857m ±  40%   -32.56% (p=0.019 n=10)
TimsortIRevSorted1M-20            2.694m ±  53%    2.820m ±  33%         ~ (p=0.971 n=10)
StandardSortIRevSorted1M-20       3.499m ±  51%    4.177m ±  33%         ~ (p=0.529 n=10)
TimsortIRandom1M-20               144.7m ±   4%    160.3m ±  33%   +10.78% (p=0.011 n=10)
StandardSortIRandom1M-20          141.1m ±   3%    150.9m ±   5%    +6.98% (p=0.001 n=10)
TimsortStrXor100-20              10.330µ ±  50%    4.853µ ±  12%   -53.02% (p=0.005 n=10)
StandardSortStrXor100-20          5.594µ ±  42%    4.127µ ±   8%   -26.23% (p=0.011 n=10)
TimsortStrSorted100-20            2.866µ ±  22%    2.988µ ±  13%         ~ (p=0.971 n=10)
StandardSortStrSorted100-20       3.726µ ±   9%    3.758µ ±  17%         ~ (p=0.796 n=10)
TimsortStrRevSorted100-20         3.906µ ±  19%    3.581µ ±  14%         ~ (p=0.247 n=10)
StandardSortStrRevSorted100-20    6.710µ ±  16%    6.691µ ±  18%         ~ (p=0.684 n=10)
TimsortStrRandom100-20            5.507µ ±  11%   11.262µ ±  43%  +104.51% (p=0.000 n=10)
StandardSortStrRandom100-20       4.413µ ±   7%    9.079µ ± 204%  +105.73% (p=0.000 n=10)
TimsortStrXor1K-20                76.02µ ±  14%    89.92µ ±  29%         ~ (p=0.143 n=10)
StandardSortStrXor1K-20           121.3µ ±  10%    128.3µ ±  90%         ~ (p=0.247 n=10)
TimsortStrSorted1K-20             49.30µ ±  44%    36.58µ ±  26%   -25.82% (p=0.043 n=10)
StandardSortStrSorted1K-20        249.3µ ± 122%    102.4µ ±  38%   -58.94% (p=0.000 n=10)
TimsortStrRevSorted1K-20         113.58µ ± 139%    30.65µ ±  13%   -73.01% (p=0.000 n=10)
StandardSortStrRevSorted1K-20    234.42µ ±  72%    99.39µ ±  15%   -57.60% (p=0.000 n=10)
TimsortStrRandom1K-20             185.4µ ±  16%    156.8µ ±  27%   -15.41% (p=0.011 n=10)
StandardSortStrRandom1K-20        180.1µ ±  26%    117.9µ ±  91%         ~ (p=0.075 n=10)
TimsortStrXor1M-20                123.4m ±   6%    123.8m ±   6%         ~ (p=0.796 n=10)
StandardSortStrXor1M-20           94.30m ±   6%    90.55m ±   7%         ~ (p=0.315 n=10)
TimsortStrSorted1M-20             66.28m ±   9%   150.91m ±  41%  +127.69% (p=0.000 n=10)
StandardSortStrSorted1M-20        179.9m ±   4%    185.6m ±  44%         ~ (p=0.052 n=10)
TimsortStrRevSorted1M-20         137.19m ±  36%    68.08m ±  29%   -50.38% (p=0.000 n=10)
StandardSortStrRevSorted1M-20     183.6m ±  48%    183.6m ±   4%         ~ (p=0.481 n=10)
TimsortStrRandom1M-20             451.8m ±  75%    454.9m ±  47%         ~ (p=0.912 n=10)
StandardSortStrRandom1M-20        320.2m ±  44%    291.6m ±  71%         ~ (p=0.247 n=10)
```