@psilva261 2025-04-17
Hi, sorry for the extremely late review and thanks for the change with the benchmark! Yeah this seems to be much better. At this point I doubt the `timsortint.go` is needed anymore though. (Although I might double-check on a server if performance varies drastically) Let me know if you want to do an update otherwise I can also pick up from here